implemented planner with moving base
Before the plannig, we have to solve IK to the target coords. But, when
we enabled both rot_also and base_also, IK is almost always get stuck
into the singular pose ... In my past implementation in silverbullet, I
solved IK using squared norm of rpy and was successful with both
rot_also and base_also. In this implementaion, I use quaternion
distance, so, maybe replacing it by rpy fix the problem.